Fix category heading on integration page layout
